[Refactor]: 챌린지그룹_유저경험치 도메인 분리
