+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/challengegroup/application/ChallengeGroupQueryService.java
@@ -6,7 +6,6 @@ import org.haedal.zzansuni.domain.challengegroup.ChallengeGroup;
 import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupImage;
 import org.haedal.zzansuni.domain.challengegroup.port.ChallengeGroupImageReader;
 import org.haedal.zzansuni.domain.challengegroup.port.ChallengeGroupReader;
-import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -38,25 +37,5 @@ public class ChallengeGroupQueryService {
         Page<ChallengeGroup> challengeGroups = challengeGroupReader.getChallengeGroupsShortsPaging(pageable, userId);
         return challengeGroups.map(ChallengeGroupModel.Info::from);
     }
-
-    @Transactional(readOnly = true)
-    public Page<ChallengeGroupModel.Ranking> getChallengeGroupRankingsPaging(Long challengeGroupId, Pageable pageable) {
-        Page<ChallengeGroupUserExp> challengeGroupUserExps
-                = challengeGroupReader.getUserExpPagingWithUserByChallengeGroupId(challengeGroupId, pageable);
-
-        // Page<ChallengeGroupUserExp>를 Page<ChallengeGroupModel.Ranking>으로 변환
-        // [rank]는 [Pageable]의 위치에 따라 계산된다.
-        return challengeGroupUserExps.map(e->{
-            int rank = challengeGroupUserExps.getNumber() * challengeGroupUserExps.getSize() + 1
-                    + challengeGroupUserExps.getContent().indexOf(e);
-            return ChallengeGroupModel.Ranking.from(e, rank);
-        });
-    }
-
-
-    public ChallengeGroupModel.Ranking getChallengeGroupRanking(Long challengeGroupId, Long id) {
-        return challengeGroupReader.getRanking(challengeGroupId, id);
-    }
-
 
 }
