+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/domain/userchallenge/port/ChallengeGroupUserExpStore.java
@@ -1,6 +1,6 @@
-package org.haedal.zzansuni.domain.challengegroup.port;
+package org.haedal.zzansuni.domain.userchallenge.port;
 
-import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
 
 public interface ChallengeGroupUserExpStore {
     ChallengeGroupUserExp store(ChallengeGroupUserExp challengeGroupUserExp);
