+++ zzansuni-api-server/app/src/main/java/org/haedal/zzansuni/infrastructure/userchallenge/ChallengeGroupUserExpRepository.java
@@ -1,6 +1,6 @@
-package org.haedal.zzansuni.infrastructure.challengegroup;
+package org.haedal.zzansuni.infrastructure.userchallenge;
 
-import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChallengeGroupUserExpRepository extends JpaRepository<ChallengeGroupUserExp, Long> {
