+++ zzansuni-api-server/app/src/test/java/org/haedal/zzansuni/domain/challengegroup/userchallenge/UserChallengeRecordServiceTest.java
@@ -13,8 +13,8 @@ import org.haedal.zzansuni.domain.userchallenge.UserChallenge;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeReader;
 import org.haedal.zzansuni.domain.userchallenge.application.UserChallengeService;
 import org.haedal.zzansuni.domain.userchallenge.port.UserChallengeStore;
-import org.haedal.zzansuni.domain.challengegroup.port.ChallengeGroupUserExpStore;
-import org.haedal.zzansuni.domain.challengegroup.ChallengeGroupUserExp;
+import org.haedal.zzansuni.domain.userchallenge.port.ChallengeGroupUserExpStore;
+import org.haedal.zzansuni.domain.userchallenge.ChallengeGroupUserExp;
 import org.haedal.zzansuni.domain.user.User;
 import org.haedal.zzansuni.domain.user.UserReader;
 import org.haedal.zzansuni.global.security.Role;
